requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cycler==0.10.0
Django==3.1.2
image==1.5.32
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
tk==0.1.0
